backend/requirements.txt
flask
flask-cors
openai
python-dotenv
PyJWT
bcrypt
